Move `CellInteract` and `CellUpdate` to `crate::model`
Since these constructs have become more prolific move them to a general,
shared `model` rather than continuing to use them from
`components::game`.
