+++ ui-rs/crates/mines_mogwai/src/components/mod.rs
@@ -1,4 +1,4 @@
 mod cell;
 pub mod game;
 
-pub use game::{board, CellInteract, CellInteractKind};
+pub use game::board;
