+++ ui-rs/crates/mines_mogwai/src/lib.rs
@@ -1,6 +1,7 @@
 mod api;
 mod app;
 mod components;
+mod model;
 mod routes;
 
 // When the `wee_alloc` feature is enabled, use `wee_alloc` as the global
