+++ ui-rs/crates/mines_mogwai/src/model.rs
@@ -1,0 +1,5 @@
+mod cell_interact;
+mod cell_update;
+
+pub use cell_interact::{CellInteract, CellInteractKind};
+pub use cell_update::CellUpdate;
