+++ ui-rs/crates/mines_mogwai/src/model/cell_update.rs
@@ -1,0 +1,10 @@
+pub enum CellUpdate {
+    All {
+        cells: Vec<Vec<String>>,
+    },
+    Single {
+        row: usize,
+        column: usize,
+        value: String,
+    },
+}
